refactor: move terminal_mode source files to TerminalMode directory
