+++ src/CMakeLists.txt
@@ -26,8 +26,8 @@ target_sources(kilo
         "${PROJECT_SOURCE_DIR}/src/Terminal/Terminal.hpp"
         "${PROJECT_SOURCE_DIR}/src/Terminal/Terminal.cpp"
 
-        "${PROJECT_SOURCE_DIR}/src/Terminal/mode/terminal_mode.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Terminal/mode/terminal_mode.cpp"
+        "${PROJECT_SOURCE_DIR}/src/TerminalMode/TerminalMode.hpp"
+        "${PROJECT_SOURCE_DIR}/src/TerminalMode/TerminalMode.cpp"
 
         "${PROJECT_SOURCE_DIR}/src/Terminal/File.hpp"
         "${PROJECT_SOURCE_DIR}/src/Terminal/File.cpp"
